tests/wpt/tests/svg/interact/scripted/defer-01.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <title>defer on script</title>
  <metadata>
      <h:link rel="help" href="https://svgwg.org/svg2-draft/interact.html#ScriptElement"/>
  </metadata>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <script>
    let script_run_status = "inline";
    let t = async_test("defer script runs later");
  </script>

  <script type="text/javascript" href="defer.js" defer="defer"></script>

  <script>

  t.step(() => {
    assert_equals(script_run_status, "inline", "script run status");
    script_run_status = "deferred";
  });

  </script>
</svg>
